+++ src/main/java/ru/fmtk/khlystov/booksaccounting/domain/Genre.java
@@ -8,7 +8,7 @@ import java.util.Objects;
 public class Genre {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private long id;
 
     @Column(nullable = false)
     private String name;
@@ -21,12 +21,12 @@ public class Genre {
         this(-1, name);
     }
 
-    public Genre(int id, String name) {
+    public Genre(long id, String name) {
         this.id = id;
         this.name = name;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
@@ -59,7 +59,7 @@ public class Genre {
     @Override
     public String toString() {
         return String.format("#%s %s",
-                (id == -1) ? "-" : Integer.toString(id),
+                (id == -1) ? "-" : Long.toString(id),
                 name);
     }
 }
