+++ src/main/java/ru/fmtk/khlystov/booksaccounting/repository/AuthorRepository.java
@@ -5,20 +5,6 @@ import ru.fmtk.khlystov.booksaccounting.domain.Author;
 
 import java.util.List;
 
-public interface AuthorRepository extends CrudRepository<Author, Integer> {
-    //    long count();
-//
-//    void insert(Author author);
-//
-//    Optional<Author> findByFullName(String firstName, String secondName);
-//
-//    Optional<Author> getById(int id);
-//
-//    Optional<Integer> getId(Author author);
-//
+public interface AuthorRepository extends CrudRepository<Author, Long> {
     List<Author> findAll();
-//
-//    boolean update(Author author);
-//
-//    boolean delete(Author author);
 }
