+++ src/main/java/ru/fmtk/khlystov/booksaccounting/repository/GenreRepository.java
@@ -6,7 +6,7 @@ import ru.fmtk.khlystov.booksaccounting.domain.Genre;
 import java.util.List;
 import java.util.Optional;
 
-public interface GenreRepository extends CrudRepository<Genre, Integer> {
+public interface GenreRepository extends CrudRepository<Genre, Long> {
     List<Genre> findAll();
 
     Optional<Genre> findAllByName(String name);
